requirements.txt
absl-py==0.6.1
aniso8601==3.0.0
args==0.1.0
astor==0.7.1
astroid==1.6.2
atari-py==0.1.1
autopep8==1.4.1
box2d-py==2.3.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
clint==0.5.1
cycler==0.10.0
Cython==0.28.5
docutils==0.14
Flask==0.12.2
Flask-JWT-Extended==3.7.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
floyd-cli==0.11.14
future==0.16.0
gast==0.2.0
glfw==1.7.0
grpcio==1.16.0
h5py==2.8.0
idna==2.6
imageio==2.4.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Kivy==1.10.1
Kivy-Garden==0.1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.0
marshmallow==2.16.2
matplotlib==2.2.2
mccabe==0.6.1
mujoco-py==1.50.1.56
numpy==1.14.3
pathlib2==2.3.2
Pillow==5.1.0
pip-save==0.2.0
ppaquette-gym-doom==0.0.6
protobuf==3.6.1
pycodestyle==2.4.0
pycparser==2.18
pygame==1.9.3
pyglet==1.3.2
Pygments==2.2.0
PyJWT==1.6.0
pylint==1.8.3
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.3
PyYAML==3.13
raven==6.9.0
requests==2.18.4
requests-toolbelt==0.8.0
scandir==1.9.0
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.2.5
tabulate==0.8.2
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
urllib3==1.22
virtualenv==16.0.0
Werkzeug==0.14.1
wrapt==1.10.11
